User Model and Routes Added

+++ routes/movies.routes.js
@@ -4,10 +4,16 @@ const Movie = require("../models/Movie.model");
 const Celebrity = require("../models/Celebrity.model");
 
 // all your routes here
+function guardRoute(req, res, next){
+  if(!req.session.currentUser){
+    res.redirect("/login");
+  }
+}
 
 //CREATE
 //Route GET to Add New Movie
 router.get("/create", (req, res, next) => {
+  guardRoute(req, res);
   Celebrity.find()
     .then((allCelebrities) => {
       res.render("movies/new-movie", {allCelebrities});
@@ -19,10 +25,12 @@ router.get("/create", (req, res, next) => {
 
 //Route POST to Add New Movie
 router.post('/create', (req, res, next) => {
+  guardRoute(req, res);
   Movie.create({
     title: req.body.movtitle,
     genre: req.body.movgenre,
     plot: req.body.movplot,
+    donor: req.session.currentUser._id
   })
   .then((dbMovie) => {
     dbMovie.cast.push(req.body.movcelebrity);
@@ -46,6 +54,9 @@ router.post('/create', (req, res, next) => {
 //READ
 //Route to display all Movies
 router.get("/", (req, res, next) => {
+  //Insert Auth
+  guardRoute(req, res);
+  console.log(req.session);
   Movie.find()
     .then((allMovies) => {
       res.render("movies/movies", { movies: allMovies });
@@ -58,9 +69,11 @@ router.get("/", (req, res, next) => {
 
 //Route to display Movie Details
 router.get("/:theId", (req, res, next) => {
-  Movie.findById(req.params.theId).populate("cast")
+  guardRoute(req, res);
+  Movie.findById(req.params.theId).populate("cast").populate("donor")
   .then((dbMovie) => {
-    res.render("movies/movie-details", dbMovie);
+    const donorMatch = dbMovie.donor.equals(req.session.currentUser._id)
+    res.render("movies/movie-details",{dbMovie, donorMatch});
   }) 
   .catch((err) => {
     next(err);
@@ -71,6 +84,7 @@ router.get("/:theId", (req, res, next) => {
 //UPDATE
 //Route GET to edit a Movie
 router.get("/:theId/edit", (req, res, next) => {
+  guardRoute(req, res);
   Movie.findById(req.params.theId).populate("cast")
   .then((theMovie)=> {
     Celebrity.find()
@@ -92,6 +106,7 @@ router.get("/:theId/edit", (req, res, next) => {
 
 //Route POST to edit a Movie
 router.post("/:theId", (req, res, next) => {
+  guardRoute(req, res);
   const {title, genre, plot, cast} = req.body; 
   Movie.findByIdAndUpdate(req.params.theId, {title, genre, plot, cast}, {new: true})
   .then((dbMovie) => {
@@ -140,9 +155,16 @@ router.post("/:theId", (req, res, next) => {
       
 //Route to delete a Movie and Cast Members with Async/Await
 router.post("/:theId/delete", async (req, res, next) => {
+  guardRoute(req, res);
 
   try {
-      const dbMovie = await Movie.findByIdAndDelete(req.params.theId).populate("cast");
+      const dbMovie = await Movie.findById(req.params.theId);
+      if(!dbMovie.donor.equals(req.session.currentUser._id)){
+        res.redirect("/");
+        return;
+      }
+      dbMovie = await Movie.findByIdAndDelete(req.params.theId).populate("cast");
+      
       const allCelebrities = await Celebrity.find(); 
 
       allCelebrities.forEach((celeb) => {
